requirements.txt
altgraph==0.17.4
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
keyboard==0.13.5
MarkupSafe==2.1.5
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dotenv==1.0.1
pytweening==1.2.0
pywin32-ctypes==0.2.2
requests==2.31.0
rsa==4.9
selenium==4.20.0
setuptools==69.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
tk==0.1.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.2
wsproto==1.2.0
